feat: Add CI pipeline
This commit introduces a CI pipeline that automatically builds and pushes docker images to Docker Hub upon pushing to the `main` branch or creating a pull request. The pipeline utilizes semantic-release for versioning and tagging.

This pipeline ensures that the latest code is always readily available in a Docker image, simplifying deployment and testing processes.

Closes #1